+++ app/src/main/AndroidManifest.xml
@@ -19,8 +19,7 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/Theme.Penguino"
-        android:usesCleartextTraffic="true"
-        tools:targetApi="33">
+        android:usesCleartextTraffic="true">
         <activity
             android:name=".ComposeActivity"
             android:exported="true"
@@ -34,18 +33,8 @@
         </activity>
 
         <service
-            android:name=".bluetooth.services.BluetoothLeService"
+            android:name=".services.BluetoothLeService"
             android:exported="false" />
-
-<!--        <activity-->
-<!--            android:name=".MainActivity"-->
-<!--            android:exported="true">-->
-<!--            <intent-filter>-->
-<!--                <action android:name="android.intent.action.MAIN" />-->
-
-<!--                <category android:name="android.intent.category.LAUNCHER" />-->
-<!--            </intent-filter>-->
-<!--        </activity>-->
     </application>
 
 </manifest>
