--- app/src/main/java/com/penguino/HomeFragment.kt
@@ -1,33 +1,0 @@
-package com.penguino
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.navigation.fragment.findNavController
-import com.penguino.databinding.FragmentHomeBinding
-
-class HomeFragment : Fragment() {
-    private lateinit var binding: FragmentHomeBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        arguments?.let {
-        }
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        binding = FragmentHomeBinding.inflate(inflater)
-
-        binding.buttonSetup.setOnClickListener {
-            findNavController().navigate(R.id.action_homeFragment_to_penguinoScanFragment)
-        }
-
-        return binding.root
-    }
-
-}
