--- app/src/main/java/com/penguino/MainActivity.kt
@@ -1,34 +1,0 @@
-package com.penguino
-
-import android.Manifest
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.util.Log
-import androidx.activity.result.ActivityResultLauncher
-import androidx.activity.result.contract.ActivityResultContracts
-
-class MainActivity : AppCompatActivity() {
-    // A popup for granting access. This is temporary and I am looking for better alternatives.
-    private val multiplePermissionLauncher: ActivityResultLauncher<Array<String>> =
-        registerForActivityResult(
-            ActivityResultContracts.RequestMultiplePermissions()
-        ) { results ->
-            results.forEach {
-                Log.d("BOOBA", "${it.key}: ${it.value}")
-            }
-        }
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-
-        // Ask permission on application launch
-        multiplePermissionLauncher.launch(arrayOf(
-            Manifest.permission.BLUETOOTH_CONNECT,
-            Manifest.permission.BLUETOOTH_SCAN,
-            Manifest.permission.ACCESS_FINE_LOCATION,
-            Manifest.permission.ACCESS_COARSE_LOCATION,
-        ))
-    }
-
-
-}
