+++ app/src/main/java/com/penguino/views/RegistrationPage.kt
@@ -14,13 +14,11 @@ import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.Button
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Snackbar
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
@@ -28,21 +26,16 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.core.text.isDigitsOnly
-import androidx.lifecycle.viewmodel.compose.LocalViewModelStoreOwner
-import androidx.lifecycle.viewmodel.compose.viewModel
-import com.penguino.bluetooth.models.RegistrationInfo
-import com.penguino.ui.theme.PenguinoTheme
-import com.penguino.utils.http.RegistrationRepository
-import com.penguino.viewmodels.RegistrationVM
+import com.penguino.models.RegistrationInfo
+import com.penguino.viewmodels.RegistrationViewModel
 
 private const val TAG = "RegistrationPage"
 @Composable
 fun RegistrationPage(
     modifier: Modifier = Modifier,
-    regVM: RegistrationVM,
+    regVM: RegistrationViewModel,
     onNavigateToRemoteControl: () -> Unit
 ) {
     Column(
@@ -82,7 +75,7 @@ fun RegistrationPage(
 @Composable
 fun TextFields(
     modifier: Modifier = Modifier,
-    regVM: RegistrationVM,
+    regVM: RegistrationViewModel,
     updater: ((RegistrationInfo) -> Unit) -> Unit
 ) {
     Column (
