--- app/src/main/res/layout/fragment_home.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:orientation="vertical"
-    android:gravity="center"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".HomeFragment">
-
-    <Button
-        android:id="@+id/button_setup"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Setup Penguino" />
-
-
-
-</LinearLayout>
