--- app/src/main/res/layout/fragment_penguino_rc.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:gravity="center"
-    tools:context=".bluetooth.fragments.PenguinoRcFragment">
-
-    <!-- TODO: Update blank fragment layout -->
-
-    <Button
-        android:id="@+id/led_on"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="ON" />
-
-    <Button
-        android:id="@+id/led_off"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="OFF" />
-
-    <Button
-        android:id="@+id/button_disconnect"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:enabled="false"
-        android:text="Disconnect" />
-</LinearLayout>
