--- app/src/main/res/layout/fragment_register_name.xml
@@ -1,47 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/textSample"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".web.fragments.RegisterNameFragment"
-    android:layout_gravity="center_horizontal"
-    android:orientation="vertical">
-
-    <EditText
-        android:id="@+id/addName"
-        android:layout_width="match_parent"
-        android:layout_height="50dp"
-        android:inputType="textPersonName"
-        android:hint="@string/edit_name"
-        android:autofillHints="" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recycle_name_suggestions"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content" />
-
-    <EditText
-        android:id="@+id/addPersonality"
-        android:layout_width="400dp"
-        android:layout_height="50dp"
-        android:inputType="text"
-        android:hint="@string/edit_personality"
-        android:importantForAutofill="no" />
-
-    <EditText
-        android:id="@+id/addAge"
-        android:layout_width="400dp"
-        android:layout_height="50dp"
-        android:inputType="number"
-        android:hint="@string/edit_age"
-        android:importantForAutofill="no" />
-
-    <Button
-        android:id="@+id/registerBtn"
-        android:layout_width="400dp"
-        android:layout_height="wrap_content"
-        android:text="@string/button_register" />
-
-</LinearLayout>
