--- app/src/main/res/layout/recycler_device_list.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="vertical"
-    android:layout_gravity="start">
-
-    <TextView
-        android:id="@+id/text_device_name"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text=""/>
-    <TextView
-        android:id="@+id/text_device_address"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="" />
-</LinearLayout>
