--- app/src/main/res/layout/recycler_suggested_name.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:overScrollMode="ifContentScrolls"
-    android:layout_gravity="center"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content">
-    <TextView
-        android:id="@+id/text_suggested_name"
-        android:layout_width="wrap_content"
-        android:layout_height="50dp"
-        android:layout_margin="3dp"
-        android:padding="8dp"
-
-        android:textColor="@color/design_default_color_on_secondary"
-        android:background="@color/design_default_color_on_primary" />
-</FrameLayout>
