--- app/src/main/res/navigation/navigation.xml
@@ -1,62 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    app:startDestination="@id/homeFragment"
-    android:id="@+id/bluetooth_navigation">
-
-    <fragment
-        android:id="@+id/homeFragment"
-        android:name="com.penguino.HomeFragment"
-        android:label="fragment_home"
-        tools:layout="@layout/fragment_home" >
-        <action
-            android:id="@+id/action_homeFragment_to_penguinoScanFragment"
-            app:destination="@id/penguinoScanFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/penguinoScanFragment"
-        android:name="com.penguino.bluetooth.fragments.PenguinoScanFragment"
-        android:label="fragment_penguino_scan"
-        tools:layout="@layout/fragment_penguino_scan" >
-        <action
-            android:id="@+id/action_penguinoScanFragment_to_devicesFoundFragment"
-            app:destination="@id/devicesFoundFragment" />
-        <action
-            android:id="@+id/action_penguinoScanFragment_to_homeFragment"
-            app:destination="@id/homeFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/devicesFoundFragment"
-        android:name="com.penguino.bluetooth.fragments.DevicesFoundFragment"
-        android:label="fragment_devices_found"
-        tools:layout="@layout/fragment_devices_found" >
-        <action
-            android:id="@+id/action_devicesFoundFragment_to_registerNameFragment"
-            app:destination="@id/registerNameFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/registerNameFragment"
-        android:name="com.penguino.web.fragments.RegisterNameFragment"
-        android:label="fragment_register_name"
-        tools:layout="@layout/fragment_register_name">
-        <argument
-            android:name="selected_device"
-            app:argType="com.penguino.bluetooth.models.DeviceInfo" />
-        <action
-            android:id="@+id/registerNameFragment_to_penguinoRcFragment"
-            app:destination="@id/penguinoRcFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/penguinoRcFragment"
-        android:name="com.penguino.bluetooth.fragments.PenguinoRcFragment"
-        android:label="fragment_penguino_rc"
-        tools:layout="@layout/fragment_penguino_rc">
-        <argument
-            android:name="selected_device"
-            app:argType="com.penguino.bluetooth.models.DeviceInfo" />
-        <action
-            android:id="@+id/action_penguinoRcFragment_to_homeFragment"
-            app:destination="@id/homeFragment" />
-    </fragment>
-</navigation>
