+++ app/src/main/res/values/strings.xml
@@ -1,7 +1,6 @@
 <resources>
     <string name="app_name">Penguino</string>
     <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="edit_personality">Personality</string>
     <string name="edit_name">Name</string>
     <string name="edit_age">Age</string>
